refactor: relocate validation_dependency_ids from runner into core hygiene
Prep for the runner extraction, applying the "single-machine core vs runner
feature" lens: a component's declared validation_dependencies come from its
portable homeboy.json manifest and are the same whether or not a runner is
present. validation_dependency_ids (+ its collector) is pure manifest parsing
mis-filed under runner.

Moved both functions into core's hygiene module (which already wrapped and was
the sole external consumer). Runner's own internal caller now uses
crate::hygiene::validation_dependency_ids; the runner re-export is dropped.

Severs one of hygiene's two core -> runner edges (copy_snapshot_to_directory,
which is woven into runner's snapshot machinery, remains for a later stage).

Verified: cargo check clean for -p homeboy-core --lib/--tests.
