+++ crates/homeboy-core/src/runner/validation_dependencies.rs
@@ -66,7 +66,7 @@ fn validation_dependency_workspaces(
     local_path: &Path,
     excludes: &[String],
 ) -> Result<Vec<PreparedValidationDependencyWorkspace>> {
-    let dependency_ids = validation_dependency_ids(local_path)?;
+    let dependency_ids = crate::hygiene::validation_dependency_ids(local_path)?;
     if dependency_ids.is_empty() {
         return Ok(Vec::new());
     }
@@ -83,57 +83,9 @@ fn validation_dependency_workspaces(
         .collect()
 }
 
-pub(crate) fn validation_dependency_ids(local_path: &Path) -> Result<Vec<String>> {
-    let manifest_path = local_path.join(PORTABLE_CONFIG_FILE);
-    let Ok(content) = fs::read_to_string(&manifest_path) else {
-        return Ok(Vec::new());
-    };
-    let manifest: serde_json::Value = serde_json::from_str(&content).map_err(|err| {
-        Error::validation_invalid_argument(
-            PORTABLE_CONFIG_FILE,
-            format!("failed to parse {}: {err}", manifest_path.display()),
-            None,
-            None,
-        )
-    })?;
-
-    let mut ids = Vec::new();
-    let Some(extensions) = manifest
-        .get("extensions")
-        .and_then(|value| value.as_object())
-    else {
-        return Ok(ids);
-    };
-
-    for extension in extensions.values() {
-        collect_validation_dependency_ids(extension, &mut ids);
-        if let Some(settings) = extension.get("settings") {
-            collect_validation_dependency_ids(settings, &mut ids);
-        }
-    }
-
-    ids.sort();
-    ids.dedup();
-    Ok(ids)
-}
-
-fn collect_validation_dependency_ids(value: &serde_json::Value, ids: &mut Vec<String>) {
-    let Some(dependencies) = value
-        .get("validation_dependencies")
-        .and_then(|value| value.as_array())
-    else {
-        return;
-    };
-
-    ids.extend(
-        dependencies
-            .iter()
-            .filter_map(|value| value.as_str())
-            .map(str::trim)
-            .filter(|value| !value.is_empty())
-            .map(str::to_string),
-    );
-}
+// validation_dependency_ids (+ its collector) is pure single-machine manifest
+// parsing; it now lives in core's hygiene module. Called here via
+// crate::hygiene::validation_dependency_ids.
 
 fn canonicalize_dependency_path(path: &Path, dependency_id: &str) -> Result<PathBuf> {
     path.canonicalize().map_err(|err| {
